Remove dead patches and other files re-introduced by merge.

+++ java-1.8.0-openjdk-remove-intree-libraries.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 
-ZIP_SRC=jdk8/jdk/src/share/native/java/util/zip/zlib-*
-JPEG_SRC=jdk8/jdk/src/share/native/sun/awt/image/jpeg
-GIF_SRC=jdk8/jdk/src/share/native/sun/awt/giflib
-PNG_SRC=jdk8/jdk/src/share/native/sun/awt/libpng
-LCMS_SRC=jdk8/jdk/src/share/native/sun/java2d/cmm/lcms
+ZIP_SRC=openjdk/jdk/src/share/native/java/util/zip/zlib-*
+JPEG_SRC=openjdk/jdk/src/share/native/sun/awt/image/jpeg/jpeg-6b
+GIF_SRC=openjdk/jdk/src/share/native/sun/awt/giflib
+PNG_SRC=openjdk/jdk/src/share/native/sun/awt/libpng
+LCMS_SRC=openjdk/jdk/src/share/native/sun/java2d/cmm/lcms
 
 echo "Removing built-in libs (they will be linked)"
 
@@ -21,17 +21,16 @@ if [ ! -f ${JPEG_SRC}/jdhuff.c ]; then # some file that sound definitely exist
 	exit 1
 fi	
 
+if false ; then
 rm -vf ${JPEG_SRC}/jcomapi.c
 rm -vf ${JPEG_SRC}/jdapimin.c
 rm -vf ${JPEG_SRC}/jdapistd.c
 rm -vf ${JPEG_SRC}/jdcoefct.c
 rm -vf ${JPEG_SRC}/jdcolor.c
-rm -vf ${JPEG_SRC}/jdct.h
 rm -vf ${JPEG_SRC}/jddctmgr.c
 rm -vf ${JPEG_SRC}/jdhuff.c
-rm -vf ${JPEG_SRC}/jdhuff.h
 rm -vf ${JPEG_SRC}/jdinput.c
-rm -vf ${JPEG_SRC}/jdmainct.c
+rm -vf ${JPEG_SRC}/jdmainct.
 rm -vf ${JPEG_SRC}/jdmarker.c
 rm -vf ${JPEG_SRC}/jdmaster.c
 rm -vf ${JPEG_SRC}/jdmerge.c
@@ -39,18 +38,12 @@ rm -vf ${JPEG_SRC}/jdphuff.c
 rm -vf ${JPEG_SRC}/jdpostct.c
 rm -vf ${JPEG_SRC}/jdsample.c
 rm -vf ${JPEG_SRC}/jerror.c
-rm -vf ${JPEG_SRC}/jerror.h
 rm -vf ${JPEG_SRC}/jidctflt.c
 rm -vf ${JPEG_SRC}/jidctfst.c
 rm -vf ${JPEG_SRC}/jidctint.c
 rm -vf ${JPEG_SRC}/jidctred.c
-rm -vf ${JPEG_SRC}/jinclude.h
 rm -vf ${JPEG_SRC}/jmemmgr.c
-rm -vf ${JPEG_SRC}/jmemsys.h
 rm -vf ${JPEG_SRC}/jmemnobs.c
-rm -vf ${JPEG_SRC}/jmorecfg.h
-rm -vf ${JPEG_SRC}/jpegint.h
-rm -vf ${JPEG_SRC}/jpeglib.h
 rm -vf ${JPEG_SRC}/jquant1.c
 rm -vf ${JPEG_SRC}/jquant2.c
 rm -vf ${JPEG_SRC}/jutils.c
@@ -60,9 +53,7 @@ rm -vf ${JPEG_SRC}/jccoefct.c
 rm -vf ${JPEG_SRC}/jccolor.c
 rm -vf ${JPEG_SRC}/jcdctmgr.c
 rm -vf ${JPEG_SRC}/jchuff.c
-rm -vf ${JPEG_SRC}/jchuff.h
 rm -vf ${JPEG_SRC}/jcinit.c
-rm -vf ${JPEG_SRC}/jconfig.h
 rm -vf ${JPEG_SRC}/jcmainct.c
 rm -vf ${JPEG_SRC}/jcmarker.c
 rm -vf ${JPEG_SRC}/jcmaster.c
@@ -75,37 +66,39 @@ rm -vf ${JPEG_SRC}/jdtrans.c
 rm -vf ${JPEG_SRC}/jfdctflt.c
 rm -vf ${JPEG_SRC}/jfdctfst.c
 rm -vf ${JPEG_SRC}/jfdctint.c
-rm -vf ${JPEG_SRC}/jversion.h
 rm -vf ${JPEG_SRC}/README
+fi
 
 echo "Removing giflib"
 if [ ! -d ${GIF_SRC} ]; then
 	echo "${GIF_SRC} does not exist. Refusing to proceed."
 	exit 1
 fi	
+if false; then
 rm -rvf ${GIF_SRC}
+fi
 
 echo "Removing libpng"
 if [ ! -d ${PNG_SRC} ]; then
 	echo "${PNG_SRC} does not exist. Refusing to proceed."
 	exit 1
 fi	
+if false; then
 rm -rvf ${PNG_SRC}
+fi
 
 echo "Removing lcms"
 if [ ! -d ${LCMS_SRC} ]; then
 	echo "${LCMS_SRC} does not exist. Refusing to proceed."
 	exit 1
 fi
-# temporary change to move bundled LCMS
-if [ ! true ]; then
+if false; then
 rm -vf ${LCMS_SRC}/cmscam02.c
 rm -vf ${LCMS_SRC}/cmscgats.c
 rm -vf ${LCMS_SRC}/cmscnvrt.c
 rm -vf ${LCMS_SRC}/cmserr.c
 rm -vf ${LCMS_SRC}/cmsgamma.c
 rm -vf ${LCMS_SRC}/cmsgmt.c
-rm -vf ${LCMS_SRC}/cmshalf.c
 rm -vf ${LCMS_SRC}/cmsintrp.c
 rm -vf ${LCMS_SRC}/cmsio0.c
 rm -vf ${LCMS_SRC}/cmsio1.c
